ai/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
ultralytics
torch
torchvision
python-multipart
httpx
openai
langchain
langchain-ollama
langchain-milvus
langchain-community
langchain-huggingface
sentence-transformers
pymilvus
python-dotenv
loguru
